# frontend-app/src/app/api/backend-core/models/dashboard-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PhysicalRiskScenario } from '../models/physical-risk-scenario';
import { TransitionRiskScenario } from '../models/transition-risk-scenario';
export interface DashboardRequest {
  physicalRiskScenario: PhysicalRiskScenario;
  transitionRiskScenario: TransitionRiskScenario;
}